refactor: extract video ID parsing logic to util function extractVideoIdFromUrl
