+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Refactored
+- Extracted video ID parsing logic from `refreshBrowsingVideos` into a dedicated utility function `extractVideoIdFromUrl` in `src/content/utils/video.ts` for better code reuse and clarity.
+
 ## [2.6.0] - 2025-07-04
 
 ### Fixed
